Moves ui method out of hook
