+++ apps/judicial-system/web/src/utils/hooks/useS3Upload/useS3Upload.ts
@@ -425,33 +425,11 @@ export const useS3Upload = (caseId: string) => {
     [formatMessage, limitedAccess, remove],
   )
 
-  const generateSingleFileUpdate = useCallback(
-    (prevFiles: TUploadFile[], displayFile: TUploadFile, newId?: string) => {
-      const index = prevFiles.findIndex((f) => f.id === displayFile.id)
-
-      if (index === -1) {
-        return prevFiles
-      }
-
-      const displayFileWithId = {
-        ...prevFiles[index],
-        ...displayFile,
-        id: newId ?? displayFile.id,
-      }
-      const next = [...prevFiles]
-      next[index] = displayFileWithId
-
-      return next
-    },
-    [],
-  )
-
   return {
     handleChange,
     handleRetry,
     handleRemove,
     uploadFromPolice,
-    generateSingleFileUpdate,
   }
 }
